pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.10.0,<4.0.0",
    "cryptography>=46.0.7",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mac-vendor-lookup>=0.1.15",
    "onvif-zeep>=0.2.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "python-nmap>=0.7.1",
    "scapy>=2.7.0",
    "smbprotocol>=1.16.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
